// packages/@aws-cdk/integ-runner/lib/workers/common.d.ts
export interface IntegTestWorkerConfig extends IntegTestInfo {
    /**
     * A list of any destructive changes
     *
     * @default []
     */
    readonly destructiveChanges?: DestructiveChange[];
}
/**
 * Information on any destructive changes
 */
export interface DestructiveChange {
    /**
     * The logicalId of the resource with a destructive change
     */
    readonly logicalId: string;
    /**
     * The name of the stack that contains the destructive change
     */
    readonly stackName: string;
    /**
     * The impact of the destructive change
     */
    readonly impact: ResourceImpact;
}
/**
 * Represents integration tests metrics for a given worker
 */
export interface IntegRunnerMetrics {
    /**
     * The region the test was run in
     */
    readonly region: string;
    /**
     * The total duration of the worker.
     * This will be the sum of all individual test durations
     */
    readonly duration: number;
    /**
     * Contains the duration of individual tests that the
     * worker executed.
     *
     * Map of testName to duration.
     */
    readonly tests: {
        [testName: string]: number;
    };
    /**
     * The profile that was used to run the test
     *
     * @default - default profile
     */
    readonly profile?: string;
}
export interface SnapshotVerificationOptions {
    /**
     * Retain failed snapshot comparisons
     *
     * @default false
     */
    readonly retain?: boolean;
    /**
     * Verbose mode
     *
     * @default false
     */
    readonly verbose?: boolean;
}
/**
 * Integration test results
 */
export interface IntegBatchResponse {
    /**
     * List of failed tests
     */
    readonly failedTests: IntegTestInfo[];
    /**
     * List of Integration test metrics. Each entry in the
     * list represents metrics from a single worker (account + region).
     */
    readonly metrics: IntegRunnerMetrics[];
}
/**
 * Common options for running integration tests
 */
export interface IntegTestOptions {
    /**
     * A list of integration tests to run
     * in this batch
     */
    readonly tests: IntegTestWorkerConfig[];
    /**
     * Whether or not to destroy the stacks at the
     * end of the test
     *
     * @default true
     */
    readonly clean?: boolean;
    /**
     * When this is set to `true` the snapshot will
     * be created _without_ running the integration test
     * The resulting snapshot SHOULD NOT be checked in
     *
     * @default false
     */
    readonly dryRun?: boolean;
    /**
     * The level of verbosity for logging.
     * Higher number means more output.
     *
     * @default 0
     */
    readonly verbosity?: number;
    /**
     * If this is set to true then the stack update workflow will be disabled
     *
     * @default true
     */
    readonly updateWorkflow?: boolean;
    /**
     * true if running in watch mode
     *
     * @default false
     */
    readonly watch?: boolean;
}